@@ -1,6 +1,6 @@
 {
   "name": "onebacklog",
-  "version": "1.18.0",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -8,19 +8,21 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/vue-fontawesome": "^2.0.6",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/vue-fontawesome": "^3.0.2",
     "axios": "^1.2.1",
     "bootstrap": "^4.6.2",
     "bootstrap-vue": "^2.23.0",
-    "popper.js": "^1.16.1",
-    "vue": "^2.7.14",
+    "@popperjs/core": "^2.11.6",
+    "core-js": "^3.26.1",
+    "mitt": "^3.0.0",
+    "vue": "^3.2.45",   
     "vue-multipane": "^0.9.5",
-    "vue-router": "^3.5.3",
+    "vue-router": "^4.1.6",
     "vue2-editor": "^2.10.3",
     "vuedraggable": "^2.24.3",
-    "vuex": "^3.6.2"
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^5.0.8",
@@ -28,12 +30,13 @@
     "@vue/cli-plugin-router": "^5.0.8",
     "@vue/cli-plugin-vuex": "^5.0.8",
     "@vue/cli-service": "^5.0.8",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^7.20.0",
+    "@vue/compat": "3.2.45",
+    "@babel/core": "^7.20.5",
+    "@babel/eslint-parser": "^7.19.1",
+    "eslint": "^8.30.0",
+    "eslint-plugin-vue": "^9.8.0",
     "node-sass": "^8.0.0",
-    "sass-loader": "^13.2.0",
-    "vue-template-compiler": "^2.7.14"
+    "sass-loader": "^13.2.0"
   },
   "eslintConfig": {
     "root": true,
@@ -42,11 +45,11 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     },
     "rules": {
       "no-console": 2,
@@ -59,6 +62,7 @@
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not dead"
+    "not dead",
+    "not ie 11"
   ]
 }
