@@ -56,7 +56,7 @@
 </template>
 
 <script>
-import AppHeader from '../header/header.vue'
+import AppHeader from '../header/Header-comp.vue'
 
 export default {
   /* Prevent accidental reloading of this page */
@@ -64,7 +64,7 @@ export default {
     window.addEventListener("beforeunload", this.preventNav)
   },
 
-  beforeDestroy() {
+  beforeUnmount() {
     window.removeEventListener("beforeunload", this.preventNav)
   },
 
