@@ -1,10 +1,9 @@
 import { SEV } from '../../../constants.js'
 import commonContext from '../common_context.js'
 import { utilities } from '../../mixins/generic.js'
-import { eventBus } from '../../../main'
 
 function created () {
-  eventBus.$on('context-menu', (node) => {
+  this.eventBus.on('context-menu', (node) => {
     this.showContextMenu(node)
   })
 }
